Rectification de la route delete

+++ backend/controllers/sauce.js
@@ -51,7 +51,17 @@ exports.modifySauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Objet supprimé" }))
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (!sauce) {
+        return res.status(404).json({ error: "Objet non trouvé" });
+      }
+      if (sauce.userId !== req.auth.userId) {
+        return res.status(401).json({ error: "Requête non autorisée" });
+      }
+      Sauce.deleteOne({ _id: req.params.id })
+        .then(() => res.status(200).json({ message: "Objet supprimé" }))
+        .catch((error) => res.status(400).json({ error }));
+    })
     .catch((error) => res.status(400).json({ error }));
 };
